% Auto-generated: do not edit by hand
\name{dacHiddenDivDownloaderButton}

\alias{dacHiddenDivDownloaderButton}

\title{HiddenDivDownloaderButton component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dacHiddenDivDownloaderButton(id=NULL, label=NULL, filename=NULL, hiddenDivData=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{filename}{Character. The basename of the csvStr file to download.}

\item{hiddenDivData}{Character. String that will be provided as a download.}
}
